+++ interface/js/index.js
@@ -25,6 +25,7 @@ let touchPosition;
 const longTouch = () => {
   const positionAsProportion = utils.getPositionAsProportion(touchPosition);
 
+  // eslint-disable-next-line no-magic-numbers
   if (positionAsProportion[0] < 0.5) {
     showModal();
   } else {
@@ -40,6 +41,7 @@ const longTouch = () => {
 const shortTouch = () => {
   const positionAsProportion = utils.getPositionAsProportion(touchPosition);
 
+  // eslint-disable-next-line no-magic-numbers
   if (positionAsProportion[0] < 0.5) {
     screen.undo();
   } else {
@@ -47,15 +49,21 @@ const shortTouch = () => {
   }
 };
 
+const handleClick = () => {
+  if (Screen.getPreference('mobileFullscreen')) {
+    utils.requestFullscreen();
+  }
+};
+
 /**
  * @param {TouchEvent} touchEvent
  */
 const touchStart = (touchEvent) => {
-  touchEvent.preventDefault(); // Prevent the browser from processing emulated mouse events
-
-  if (Screen.getPreference('mobileFullscreen')) {
-    utils.requestFullscreen();
-  }
+  // FIXME: When this is enabled, the click event doesn't fire
+  // but requestFullscreen is unreliable when called from touchstart #12
+  // so we need to use the click event instead
+  // https://stackoverflow.com/questions/79132797/requestfullscreen-is-unreliable-when-called-from-touchstart
+  // touchEvent.preventDefault(); // Prevent the browser from processing emulated mouse events
 
   const touch = touchEvent.targetTouches.item(0);
   touchPosition = [touch.clientX, touch.clientY];
@@ -101,6 +109,7 @@ const touchCancel = () => {
 if (utils.mobileOrTabletCheck()) {
   const touchCatcherElement = document.getElementById('touchCatcher');
   touchCatcherElement.style.display = 'block';
+  window.addEventListener('click', handleClick, false);
   touchCatcherElement.addEventListener('touchstart', touchStart, false);
   touchCatcherElement.addEventListener('touchmove', touchMove, false);
   touchCatcherElement.addEventListener('touchcancel', touchCancel, false);
@@ -115,3 +124,16 @@ window.addEventListener('click', (event) => {
     hideModal();
   }
 });
+
+document.getElementById('backToEditorButton').addEventListener('click', () => {
+  window.history.back();
+});
+
+document
+  .getElementById('fullscreenToggle')
+  .addEventListener('click', (event) => {
+    event.target.classlist.toggle('checked');
+
+    const checked = event.target.classList.contains('checked');
+    Screen.setPreference('mobileFullscreen', checked);
+  });